iot-bigdata-pipeline/requirements.txt
pymongo==4.5.0
python-dotenv==1.0.0
pytest==7.4.0
faker==19.0.0